perf(html): inline small external CSS
